+++ lib/rapns/notification.rb
@@ -29,15 +29,6 @@ module Rapns
 
     scope :completed, lambda { where("delivered = ? OR failed = ?", true, true) }
 
-    def initialize(*args)
-      attributes = args.first
-      if attributes.is_a?(Hash) && attributes.keys.include?(:attributes_for_device)
-        msg = ":attributes_for_device via mass-assignment is deprecated. Use :data or the attributes_for_device= instance method."
-        Rapns::Deprecation.warn(msg)
-      end
-      super
-    end
-
     def data=(attrs)
       return unless attrs
       raise ArgumentError, "must be a Hash" if !attrs.is_a?(Hash)